Fix bug where parameters are ignored in new sessions
The controller was dropping attributes set in the Sessions POST, which
caused the default expiration time to always be used instead of the time
specified by the caller.

This fixes that and adds controller tests for it.  There are already
Account tests but they don't go through the controllers, hence why it
was missed.

A security concern with this fix is ensuring that the caller can't
specify their user ID and create sessions for a different user.  There
is also a new test for that scenario.

Fixes #6