During database bootstrap, promote root user to admin
The root user is there in a new database to fix a chicken/egg problem.
Only an admin can create admin users.  Regular users can't make themselves admins.

We create a default root user therefore to use for creating the "real"
admin accounts.

Speaking of this root user, make sure to either delete it or change it's
password before going to production!!!

Fixes #8