AuthController add functions for extracting IDs from conn
It's useful to take a conn and get the authed user ID and the authed
session ID out of it.  One can usually do:

```
conn.assigns.authed_user_id
conn.assigns.authed_session_id
```

Except that if the assigns or the authed user ID or session ID is nil,
then that code bombs.  For that reason it's nice to have a safer
version.  This commit adds functions for that to the Auth Controller