Add defp_testable macro for testing private functions
Generally speaking testing private functions is not recommended, but in
the case of schema files there are a bunch of small utility functions
that make absolutely no sense outside of the module, but still benefit
from testing as they are part of a long chain where it's very difficult
to exercise all code paths in a test using only the public entrypoint.

The previous solution of exporting the functions if Mix.env() == "test"
was working until newer versions of elixir.  The tests still work
properly, but elixir prints verbose warnings for every invocation.
Unfortunately this pollutes the compiler output with garbage and leads
to the ignoring of warnings due to low signal to noise ratio.

To fix this, we define a macro called `defp_testable` which will define
the function as a `defp` in non test environments, but with `def` in
a test environment.  This keeps the tests working and silences the
warning messages from the compiler