Add pagination to malan
This comes with a schema (not stored in the database but use as a way
validate data).  This also comes with a pagination module to act as
a bounded context, and a pagination controller to act on conn objects
and support other controllers that need pagination for some of their
endpoint.

Also include tests for all of it.

Refs #50