Store a copy of the changeset related to the transaction
You can usually infer the impact that a transaction had using the
existing params, but in some cases (particularly errors) it is useful to
have a copy of the actual changeset that was involved.

This changeset is stored in a jsonb column in the table.  To save space,
the entire changeset is not stored.  We only store the useful subset of
properties.  We also change the structure of the changeset a bit because
it is useful to have a noticeably different struct definition so that
there isn't any ambiguity in what object you are dealing with.

Fixes #42
Fixes #75