Remove real_ip_address column from sessions
This implementation has a benefit in that we record the user's real IP
as well as the IP of the connecting thing, but in exchange we get
a small amount of complexity.  As we add rate limiting, this complexity
doesn't seem worth it.

This patch removes the real_ip_address column and the Cloudflare header
parsing.  The state of the app at this point is that conn.remote_ip
stays default behavior.  Future commit will add back in Cloudflare
handling.

Refs #41
Refs #60