Add authorized_ips to user model
If set, these can limit a user's login based on their IP address.  If no
approved IP addresses are set, then IP will be considered/limited for
a login (although it will still be logged in the session and
associated transactions).

Note that if you are behind a proxy that isn't cloudflare (since
cloudflare is covered by plug_cloudflare) the IP address of the peer
will be the proxy and not the user, so this will fail.

Fixes #77