Add remote IP to Transactions
This way we can see the IP address of the peer who caused the
Transaction to be recorded.

As part of this, we also refactor the accounts test to use string keys
instead of atom keys.  This is something we've needed anyway, but the
changes here forced it to happen unless we wanted to put in some gross
hacks.

Fixes #87