Make map_to_string work recursively, and add list_to_string
The utility function to convert a map to a string previously didn't
recursively convert.  For example a nested map was not converted (it
would just bomb because map doesn't implement `String.Chars` protocol).

This makes recursive work.  Also adds a `list_to_string` function so
that we can support nested lists