Add env var to disable TLS in release mode
When running malan locally, sometimes we use the production/release
build.  In that situation we often don't have TLS set up locally, so
wish to disable it.

This adds an env var "DATABASE_TLS_ENABLED" which can be explicitly set
to false in order to disable TLS.  If absent, the default is true
(require TLS)

Fixes #104