Change birthday type in schema to date rather than datetime
The original idea was that "birth" can occur at specific times,
so a datetime was a good type. However in practice that extra
precision can complicate things when timezones come into play.
To avoid this problem, it would be best to have birthday be a
date instead of a datetime

Note:  For actually changing the data in postgres, this migration relies
on postgres automatic date conversion.  It was tested with a broad range
of datetimes, and each time the hours/minutes/seconds portion was
dropped and the date itself was kept by postgres.

Fixes #107