Make a few provisioning commands idempotent for robustness
Also add a bunch of nodes regarding provisioning.  Was hoping to fully
automate it with this patch but there are still a few things that won't
automate and this thing went from "Real quick" to 4 hours, so I'm
punting on that for later.  At least with this it can be done entirely
with the CLI and don't have to open the aws console