When deleting a user, revoke all active sessions
When a user is deleted, all of their active sessions should be
automatically revoked.  The session isn't usable for anything, but
should be revoked nonetheless for completeness.

Adds a handful of tests around this functionality.  Updates some tests
that were impacted.  Adds a test for access control around user deletion
to ensure that only admins or the user themselves can delete the user.

Fixes #112