Fix bug preventing multiple deletes of user
when a user is deleted, we previously prepended a sentinel value to the
username/email so they could be reused later.  However this only gave us
one delete.  if a user that had been recreated was deleted again, it
would fail a uniqueness constraint.

This patch prepends a UUID to it, so we can delete as many times as we
need to.

Fixes #114