Fix up flaky tests
Rewrite the pagination tests to be permissive on the order in which they
are returned.  This fixes a fragility in pagination tests.

- The flakiness is a result of the tests expecting a deterministic order for objects created
- Running tests serially (synchronously) makes no difference
- Running tests against a completely fresh DB instance makes no difference
- Postgres is not deterministic in this. That's not a problem, just an implementation reality

The fix is to rewrite the tests in a way that the lists are sorted
before comparison.  The net result is that order doesn't matter but
membership (and length) does:

Fixes #117