Increase pool size for prod instances from 5 to 7
with 4 instances currently we are using 20 connections.  There are 47 / 97
connections still available during peak traffic.  This would increase us
from 20 connections to 28, leaving 39 connections still unused.  I think
it should be pretty safe and still leaves us with a lot of overhead in case
of traffic bursts on other applications sharing the db