Add API for extending a session
While a session is still active, it can be extended up to the limit
specified at session creation time.  Extensions can be done using the
existing token by calling one of the new endpoints:

```
PUT /sessions/current/extend
PUT /users/:id/sessions/:id/extend
```

You can omit the body, or include it with specification of how long to
extend the session for.  Example, this body in a
`PUT /sessions/current/extend`:

```json
{
  "extend_by_seconds": 300
}
```

Extensions are logged as JSON objects in the Session object for auditing
purposes.  These are returned as part of the show (GET) endpoint for
Sessions

GET /sessions/current
GET /users/:id/sessions/:id

Some of the tests have been stubbed but are not yet completed.

New parameters in the Session object (these show up in the GET responses
now):

`extendable_until` (datetime) - UTC DateTime of the last point in time
that a session can be alive.  You can't extend past this point.  This is
configured in the create (POST) action that first created the session.
This is set through the session parameter in the JSON called
`extendable_until_seconds`.  If not explicitly set, a default value will
be used.

`max_extension_secs` (integer) - This is the farthest into the future an
extension can go.  This is also configured at session creation time
using the "max_extension_secs" field in the session JSON.  If left
empty, a default value will be used.

There is also a new object SessionExtension.  These provide a log of all
extensions made to a session.  To retrieve these, use this endpoint with
session ID:

GET /sessions/:session_id/extensions

Or to retrieve a single extension by ID:

GET /session_extensions/:id